requirements.txt
aioredis==1.3.1
altgraph==0.17.2
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.4
astroid==2.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.4
Avatar==0.1
avatar2==1.4.6
awsebcli==3.20.3
bcrypt==3.2.2
bitstring==3.1.9
blessed==1.19.1
bootstrap4==0.1.0
botocore==1.23.54
cached-property==1.5.2
capstone==4.0.2
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.5
channels==3.0.3
channels-redis==3.3.0
chardet==4.0.0
colorama==0.4.3
configparser==5.0.2
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
distlib==0.3.0
dj-database-url==0.5.0
dj-rest-auth==2.1.5
Django==3.2.3
django-admin-interface==0.16.3
django-allauth==0.44.0
django-appconf==1.0.4
django-avatar==5.0.0
django-cleanup==5.2.0
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-extensions==3.1.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-mathfilters==1.0.0
django-nested-admin==3.3.3
django-phonenumber-field==5.2.0
django-rest-multiple-models==2.1.3
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
dodgy==0.2.1
drf-writable-nested==0.6.2
filelock==3.0.12
flake8==3.8.4
flake8-polyfill==1.0.2
future==0.16.0
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
intervaltree==3.1.0
isort==4.3.21
jdatetime==3.6.2
jmespath==0.10.0
jsonschema==3.2.0
keystone-engine==0.9.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.20
npyscreen==4.10.5
oauthlib==3.1.1
paramiko==2.10.4
parse==1.19.0
pathspec==0.9.0
pep8-naming==0.10.0
phonenumbers==8.12.31
Pillow==8.2.0
pipenv==2020.5.28
posix-ipc==1.0.5
prospector==1.3.1
protobuf==3.12.2
psutil==5.8.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
pygdbmi==0.10.0.1
PyJWT==1.7.1
pylink-square==0.10.1
pylint==2.5.3
pylint-celery==0.3
pylint-flask==0.6
pylint-plugin-utils==0.6
PyNaCl==1.5.0
pyOpenSSL==20.0.1
PyPDF2==1.26.0
pyrsistent==0.18.1
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2020.1
pyusb==1.2.1
PyYAML==5.3.1
qrcode==6.1
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.0
requirements-detector==0.7
semantic-version==2.8.5
service-identity==21.1.0
setoptconf==0.2.0
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.4.0
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.4
toml==0.10.1
Twisted==21.2.0
txaio==21.2.1
unicorn==1.0.3
Unidecode==1.3.2
Unipath==1.1
urllib3==1.26.5
virtualenv==20.0.21
virtualenv-clone==0.5.4
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.59.0
wrapt==1.12.1
xhtml2pdf==0.2.5
zope.interface==5.4.0
